プレイヤーの調整

+++ Project/Player.h
@@ -17,21 +17,21 @@ public:
 	void		Finalize		 () override;		//後処理
 	void		Update			 () override;		//更新
 	const void	DrawCharacterInfo()const override;	//描画
-	void PlayAnimation(const int _nextAnimation, const float _playTime);
-	void DeathProcess();
-	const bool CanAction		(const float _staminaConsumed)const;//アクションができるか
-	void CalcStamina(const float _staminaConsumed, const float _maxStamina);			//スタミナの回復処理
+	void		PlayAnimation	 (const int _nextAnimation, const float _playTime);
+	void		DeathProcess	 ();
+	const bool	CanAction		 (const float _staminaConsumed)const;//アクションができるか
+	void		CalcStamina		 (const float _staminaConsumed, const float _maxStamina);			//スタミナの回復処理
 
 	/*getter*/
 	const int		GetNowState			()const;
 	const bool		GetIsAttack			()const override { return false; }												//攻撃フラグの取得
-	const int		GetStamina			()const;																		//スタミナの取得
+	const float		GetStamina			()const;																		//スタミナの取得
 	const int		GetHealCount		()const { return this->healCount; }												//回復オーブの数(今は回復回数になっている)
 	CharacterData&	GetPlayerData		();																				//プレイヤーデータの取得
 	Rigidbody&		GetPlayerRigidbody	();																				//リジッドボディの取得
 	const bool		GetIsDrawSword		()const { return this->isDrawSword; }											//武器を出しているか
 		  void		SetHealCount		(const int _count) { this->healCount = _count; }								//回復回数の設定
-		  void		SetHitStop			(const float _time, const int _type, const float _delay, const float _factor);	//ヒットストップの設定
+		  void		SetHitStop			(const int _time, const int _type, const int _delay, const float _factor);		//ヒットストップの設定
 		  void		SetIsDrawSword		(const bool _set) { this->isDrawSword = _set; }									//抜刀フラグの設定
 	const bool		GetIsLock			() { return this->isLock; }//ロックオンフラグ
 	//アニメーションの種類
